Fixing issues with json2struct datatransfrom from file
